services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API services"
authors = ["Yoganandan <p.yoganandan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg = "^3.1.18"
google-api-python-client = "^2.121.0"
redis = "^5.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
